Added support for thresholds with decimals

+++ src/main/java/com/vectorcast/plugins/vectorcastcoverage/VectorCASTHealthReportThresholds.java
@@ -9,24 +9,24 @@ import java.io.Serializable;
  * @since 1.7
  */
 public class VectorCASTHealthReportThresholds implements Serializable {
-    private int minStatement;
-    private int maxStatement;
-    private int minBranch;
-    private int maxBranch;
-    private int minBasisPath;
-    private int maxBasisPath;
-    private int minMCDC;
-    private int maxMCDC;
-    private int minFunction;
-    private int maxFunction;
+    private float minStatement;
+    private float maxStatement;
+    private float minBranch;
+    private float maxBranch;
+    private float minBasisPath;
+    private float maxBasisPath;
+    private float minMCDC;
+    private float maxMCDC;
+    private float minFunction;
+    private float maxFunction;
 
-    private int minFunctionCall;
-    private int maxFunctionCall;
+    private float minFunctionCall;
+    private float maxFunctionCall;
 
     public VectorCASTHealthReportThresholds() {
     }
 
-    public VectorCASTHealthReportThresholds(int minStatement, int maxStatement, int minBranch, int maxBranch, int minBasisPath, int maxBasisPath, int minMCDC, int maxMCDC, int minFunction, int maxFunction, int minFunctionCall, int maxFunctionCall) {
+    public VectorCASTHealthReportThresholds(float minStatement, float maxStatement, float minBranch, float maxBranch, float minBasisPath, float maxBasisPath, float minMCDC, float maxMCDC, float minFunction, float maxFunction, float minFunctionCall, float maxFunctionCall) {
         this.minStatement = minStatement;
         this.maxStatement = maxStatement;
         this.minBranch = minBranch;
@@ -42,7 +42,7 @@ public class VectorCASTHealthReportThresholds implements Serializable {
         ensureValid();
     }
 
-    private int applyRange(int min , int value, int max) {
+    private float applyRange(float min , float value, float max) {
         if (value < min) return min;
         if (value > max) return max;
         return value;
@@ -63,99 +63,99 @@ public class VectorCASTHealthReportThresholds implements Serializable {
         minFunctionCall = applyRange(0, minFunctionCall, maxFunctionCall);
     }
 
-    public int getMinStatement() {
+    public float getMinStatement() {
         return minStatement;
     }
 
-    public void setMinStatement(int minStatement) {
+    public void setMinStatement(float minStatement) {
         this.minStatement = minStatement;
     }
 
-    public int getMaxStatement() {
+    public float getMaxStatement() {
         return maxStatement;
     }
 
-    public void setMaxStatement(int maxStatement) {
+    public void setMaxStatement(float maxStatement) {
         this.maxStatement = maxStatement;
     }
 
-    public int getMinBranch() {
+    public float getMinBranch() {
         return minBranch;
     }
 
-    public void setMinBranch(int minBranch) {
+    public void setMinBranch(float minBranch) {
         this.minBranch = minBranch;
     }
 
-    public int getMaxBranch() {
+    public float getMaxBranch() {
         return maxBranch;
     }
 
-    public void setMaxBranch(int maxBranch) {
+    public void setMaxBranch(float maxBranch) {
         this.maxBranch = maxBranch;
     }
 
-    public int getMinBasisPath() {
+    public float getMinBasisPath() {
         return minBasisPath;
     }
 
-    public void setMinBasisPath(int minBasisPath) {
+    public void setMinBasisPath(float minBasisPath) {
         this.minBasisPath = minBasisPath;
     }
 
-    public int getMaxBasisPath() {
+    public float getMaxBasisPath() {
         return maxBasisPath;
     }
 
-    public void setMaxBasisPath(int maxBasisPath) {
+    public void setMaxBasisPath(float maxBasisPath) {
         this.maxBasisPath = maxBasisPath;
     }
 
-    public int getMinMCDC() {
+    public float getMinMCDC() {
         return minMCDC;
     }
 
-    public void setMinMCDC(int minMCDC) {
+    public void setMinMCDC(float minMCDC) {
         this.minMCDC = minMCDC;
     }
 
-    public int getMaxMCDC() {
+    public float getMaxMCDC() {
         return maxMCDC;
     }
 
-    public void setMaxMCDC(int maxMCDC) {
+    public void setMaxMCDC(float maxMCDC) {
         this.maxMCDC = maxMCDC;
     }
     
-    public void setMinFunction(int minFunction) {
+    public void setMinFunction(float minFunction) {
         this.minFunction = minFunction;
     }
 
-    public int getMinFunction() {
+    public float getMinFunction() {
         return minFunction;
     }
 
-    public void setMaxFunction(int maxFunction) {
+    public void setMaxFunction(float maxFunction) {
         this.maxFunction = maxFunction;
     }
 
-    public int getMaxFunction() {
+    public float getMaxFunction() {
         return maxFunction;
     }
     
-    public void setMinFunctionCall(int minFunctionCall) {
+    public void setMinFunctionCall(float minFunctionCall) {
         this.minFunctionCall = minFunctionCall;
     }
 
-    public int getMinFunctionCall() {
+    public float getMinFunctionCall() {
         return minFunctionCall;
     }
 
-    public void setMaxFunctionCall(int maxFunctionCall) {
+    public void setMaxFunctionCall(float maxFunctionCall) {
         this.maxFunctionCall = maxFunctionCall;
     }
 
-    public int getMaxFunctionCall() {
+    public float getMaxFunctionCall() {
         return maxFunctionCall;
     }
